+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - AutoUnplacer now prefers non-UpToDate resources after AP-settings-violating resources before choosing any.
+- Improved error message for shrink operations
 
 ### Fixed